+++ snuba_sdk/formula.py
@@ -2,12 +2,12 @@ from __future__ import annotations
 
 from dataclasses import dataclass, replace
 from enum import Enum
-from typing import Optional, Sequence, Union, Any
+from typing import Any, Optional, Sequence, Union
 
 from snuba_sdk.aliased_expression import AliasedExpression
 from snuba_sdk.column import Column
 from snuba_sdk.conditions import BooleanCondition, Condition, ConditionGroup
-from snuba_sdk.expressions import Expression, InvalidExpressionError, list_type
+from snuba_sdk.expressions import InvalidExpressionError, list_type
 from snuba_sdk.timeseries import Timeseries
 
 
@@ -29,17 +29,15 @@ class ArithmeticOperator(Enum):
 @dataclass(frozen=True)
 class Formula:
     operator: ArithmeticOperator
-    parameters: Optional[
-        Sequence[
-            Union[Formula, Timeseries, float, int]
-        ]
-    ] = None
+    parameters: Optional[Sequence[FormulaParameterGroup]] = None
     filters: Optional[ConditionGroup] = None
     groupby: Optional[list[Column | AliasedExpression]] = None
 
     def validate(self) -> None:
         if not isinstance(self.operator, ArithmeticOperator):
-            raise InvalidFormulaError(f"formula '{self.operator}' must be a ArithmeticOperator")
+            raise InvalidFormulaError(
+                f"formula '{self.operator}' must be a ArithmeticOperator"
+            )
         if self.parameters is not None:
             if not isinstance(self.parameters, Sequence):
                 raise InvalidFormulaError(
@@ -55,6 +53,15 @@ class Formula:
         new = replace(self, **{field: value})
         return new
 
+    def set_parameters(self, parameters: Sequence[FormulaParameterGroup]) -> Formula:
+        if parameters is not None and not list_type(
+            parameters, (Formula, Timeseries, float, int)
+        ):
+            raise InvalidFormulaError(
+                "parameters must be a list of either Formulas, Timeseries, floats, or ints"
+            )
+        return self._replace("parameters", parameters)
+
     def set_filters(self, filters: ConditionGroup | None) -> Formula:
         if filters is not None and not list_type(
             filters, (BooleanCondition, Condition)
@@ -62,9 +69,10 @@ class Formula:
             raise InvalidFormulaError("filters must be a list of Conditions")
         return self._replace("filters", filters)
 
-    def set_groupby(
-        self, groupby: list[Column | AliasedExpression] | None
-    ) -> Formula:
+    def set_groupby(self, groupby: list[Column | AliasedExpression] | None) -> Formula:
         if groupby is not None and not list_type(groupby, (Column, AliasedExpression)):
             raise InvalidFormulaError("groupby must be a list of Columns")
         return self._replace("groupby", groupby)
+
+
+FormulaParameterGroup = Union[Formula, Timeseries, float, int]