+++ snuba_sdk/metrics_query_visitors.py
@@ -2,15 +2,13 @@ from __future__ import annotations
 
 from abc import ABC, abstractmethod
 from datetime import datetime
-from typing import Generic, Mapping, Sequence, TypeVar
+from typing import Generic, Mapping, TypeVar
 
 # Import the module due to sphinx autodoc problems
 # https://github.com/agronholm/sphinx-autodoc-typehints#dealing-with-circular-imports
 from snuba_sdk import metrics_query as main
-from snuba_sdk.aliased_expression import AliasedExpression
 from snuba_sdk.column import Column
-from snuba_sdk.conditions import BooleanCondition, Condition, ConditionGroup, Op
-from snuba_sdk.expressions import list_type
+from snuba_sdk.conditions import Condition, Op
 from snuba_sdk.formula import Formula
 from snuba_sdk.metrics_visitors import (
     RollupSnQLPrinter,
@@ -47,16 +45,6 @@ class MetricsQueryVisitor(ABC, Generic[QVisited]):
 
     @abstractmethod
     def _visit_query(self, query: Timeseries | None) -> Mapping[str, QVisited]:
-        raise NotImplementedError
-
-    @abstractmethod
-    def _visit_filters(self, filters: ConditionGroup | None) -> QVisited:
-        raise NotImplementedError
-
-    @abstractmethod
-    def _visit_groupby(
-        self, groupby: Sequence[Column | AliasedExpression] | None
-    ) -> QVisited:
         raise NotImplementedError
 
     @abstractmethod
@@ -121,14 +109,10 @@ class SnQLPrinter(MetricsQueryVisitor[str]):
 
         if timeseries_data["groupby"]:
             groupby_columns.append(timeseries_data["groupby"])
-        if returns["groupby"]:
-            groupby_columns.append(returns["groupby"])
 
         where_clauses.append(timeseries_data["metric_filter"])
         if timeseries_data["filters"]:
             where_clauses.append(timeseries_data["filters"])
-        if returns["filters"]:
-            where_clauses.append(returns["filters"])
 
         where_clauses.append(returns["scope"])
         where_clauses.append(returns["start"])
@@ -160,23 +144,14 @@ class SnQLPrinter(MetricsQueryVisitor[str]):
             ]
         ).strip()
 
-    def _visit_query(self, query: Timeseries | None) -> Mapping[str, str]:
+    def _visit_query(self, query: Timeseries | Formula | None) -> Mapping[str, str]:
         if query is None:
             raise InvalidMetricsQueryError("MetricQuery.query must not be None")
-
+        if isinstance(query, Formula):
+            raise InvalidMetricsQueryError(
+                "Serializing a Formula in MetricQuery.query is unsupported"
+            )
         return self.timeseries_visitor.visit(query)
-
-    def _visit_filters(self, filters: ConditionGroup | None) -> str:
-        if filters is not None:
-            return f"{' AND '.join(self.expression_visitor.visit(c) for c in filters)}"
-        return ""
-
-    def _visit_groupby(
-        self, groupby: Sequence[Column | AliasedExpression] | None
-    ) -> str:
-        if groupby is not None:
-            return f"{', '.join(self.expression_visitor.visit(g) for g in groupby)}"
-        return ""
 
     def _visit_start(self, start: datetime | None) -> str:
         if start is None:
@@ -218,21 +193,6 @@ class Validator(MetricsQueryVisitor[None]):
             raise InvalidMetricsQueryError("query must be a Timeseries or Formula")
         query.validate()
         return {}  # Necessary for typing
-
-    def _visit_filters(self, filters: ConditionGroup | None) -> None:
-        if filters is not None:
-            if not list_type(filters, (Condition, BooleanCondition)):
-                raise InvalidMetricsQueryError("filters must be a list of Conditions")
-            (c.validate() for c in filters)
-
-    def _visit_groupby(
-        self, groupby: Sequence[Column | AliasedExpression] | None
-    ) -> None:
-        if groupby is not None:
-            if not list_type(groupby, (Column, AliasedExpression)):
-                raise InvalidMetricsQueryError("groupby must be a list of Columns")
-            for g in groupby:
-                g.validate()
 
     def _visit_start(self, start: datetime | None) -> None:
         if start is None: